# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Recommendations API
 * OpenAPI spec version: 0.1.0
 */
export interface GetSkillsRequest {
  careerGoal: string;
}

export interface SkillRequirement {
  name: string;
  requiredLevel: number;
  description: string;
  category?: string;
}

export interface GetSkillsResponse {
  careerGoal: string;
  skills: SkillRequirement[];
}

export interface CurrentSkill {
  name: string;
  currentLevel: number;
  requiredLevel: number;
}

export interface SkillGapRequest {
  careerGoal: string;
  currentSkills: CurrentSkill[];
}

export interface SkillGapItem {
  name: string;
  currentLevel: number;
  requiredLevel: number;
  gap: number;
  priority: string;
  resources?: string[];
}

export interface LearningPhase {
  phase: number;
  title: string;
  duration: string;
  skills: string[];
  actions: string[];
}

export interface SkillGapAnalysis {
  careerGoal: string;
  overallReadiness: number;
  estimatedTime: string;
  skills: SkillGapItem[];
  roadmap: LearningPhase[];
}

export interface HealthStatus {
  status: string;
}

export interface CareerQuizAnswer {
  questionId: string;
  answer: string;
}

export interface CareerQuizAnswers {
  answers: CareerQuizAnswer[];
}

export interface RoadmapStep {
  phase: string;
  title: string;
  description: string;
  duration: string;
  skills: string[];
  resources: string[];
}

export interface CareerRecommendation {
  careerTitle: string;
  matchScore: number;
  description: string;
  whyGoodFit: string;
  salaryRange: string;
  jobOutlook: string;
  keySkills: string[];
  alternativeCareers: string[];
  roadmap: RoadmapStep[];
}

export type CareerQuestionOptionsItem = {
  value: string;
  label: string;
};

export interface CareerQuestion {
  id: string;
  question: string;
  category: string;
  options: CareerQuestionOptionsItem[];
}

export interface CareerQuestionsResponse {
  questions: CareerQuestion[];
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};