# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Career Recommendations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skill-gap/get-skills:
    post:
      operationId: getSkillsForCareer
      tags: [career]
      summary: Get required skills for a career goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSkillsRequest"
      responses:
        "200":
          description: Skills required for the career
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSkillsResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /skill-gap/analyze:
    post:
      operationId: analyzeSkillGap
      tags: [career]
      summary: Analyze skill gap and generate learning roadmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillGapRequest"
      responses:
        "200":
          description: Skill gap analysis with roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGapAnalysis"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /career/recommend:
    post:
      operationId: getCareerRecommendation
      tags: [career]
      summary: Get career recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerQuizAnswers"
      responses:
        "200":
          description: Career recommendation with roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerRecommendation"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    GetSkillsRequest:
      type: object
      required: [careerGoal]
      properties:
        careerGoal:
          type: string
    GetSkillsResponse:
      type: object
      required: [careerGoal, skills]
      properties:
        careerGoal:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillRequirement"
    SkillRequirement:
      type: object
      required: [name, requiredLevel, description]
      properties:
        name:
          type: string
        requiredLevel:
          type: number
        description:
          type: string
        category:
          type: string
    SkillGapRequest:
      type: object
      required: [careerGoal, currentSkills]
      properties:
        careerGoal:
          type: string
        currentSkills:
          type: array
          items:
            $ref: "#/components/schemas/CurrentSkill"
    CurrentSkill:
      type: object
      required: [name, currentLevel, requiredLevel]
      properties:
        name:
          type: string
        currentLevel:
          type: number
        requiredLevel:
          type: number
    SkillGapAnalysis:
      type: object
      required: [careerGoal, overallReadiness, skills, roadmap, estimatedTime]
      properties:
        careerGoal:
          type: string
        overallReadiness:
          type: number
        estimatedTime:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillGapItem"
        roadmap:
          type: array
          items:
            $ref: "#/components/schemas/LearningPhase"
    SkillGapItem:
      type: object
      required: [name, currentLevel, requiredLevel, gap, priority]
      properties:
        name:
          type: string
        currentLevel:
          type: number
        requiredLevel:
          type: number
        gap:
          type: number
        priority:
          type: string
        resources:
          type: array
          items:
            type: string
    LearningPhase:
      type: object
      required: [phase, title, duration, skills, actions]
      properties:
        phase:
          type: number
        title:
          type: string
        duration:
          type: string
        skills:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: string
...
    CareerQuizAnswer:
      type: object
      properties:
        questionId:
          type: string
        answer:
          type: string
      required:
        - questionId
        - answer
    CareerQuizAnswers:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/CareerQuizAnswer"
      required:
        - answers
    RoadmapStep:
      type: object
      properties:
        phase:
          type: string
        title:
          type: string
        description:
          type: string
        duration:
          type: string
        skills:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            type: string
      required:
        - phase
        - title
        - description
        - duration
        - skills
        - resources
    CareerRecommendation:
      type: object
      properties:
        careerTitle:
          type: string
        matchScore:
          type: number
        matchScoreBreakdown:
          type: string
        description:
          type: string
        whyGoodFit:
          type: string
        salaryRange:
          type: string
        jobOutlook:
          type: string
        keySkills:
          type: array
          items:
            type: string
        alternativeCareers:
          type: array
          items:
            type: string
        roadmap:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapStep"
      required:
        - careerTitle
        - matchScore
        - matchScoreBreakdown
        - description
        - whyGoodFit
        - salaryRange
        - jobOutlook
        - keySkills
        - alternativeCareers
        - roadmap
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...