# shared/api-zod/src/generated/types/skillGapRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Recommendations API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentSkill } from "./currentSkill";

export interface SkillGapRequest {
  careerGoal: string;
  currentSkills: CurrentSkill[];
}